+++ Relation.cpp
@@ -16,7 +16,7 @@ void Relation::addTuple(Tuple toAdd) {
 
 void Relation::toString() {
     for (Tuple t : relations) {
-        for (int i = 0; i < relationHeader.attributes.size(); i++) {
+        for (long unsigned int i = 0; i < relationHeader.attributes.size(); i++) {
             if (i < relationHeader.attributes.size()-1) {
                 std::cout << relationHeader.attributes[i] << "=" << t.values[i] << ",";
             }
